feat(mvp): Add observable window (fix #14, #13, #11)
* feat(display-window): Set observable window proportional to canvas dimensions (closes #11)

The observable window is the section of CGOL to be displayed in the canvas, and its dimensions are proportional to the width and height of the canvas.

* feat(display-window): Set zoom minimum (closes #13)

Zoom minimum is determined by leading to a minimum size on the smallest axis (width or height). As of right now, the minimum size for the smallest axis is 64px.

* feat(window-display): Add constrains to panning (closes #14)

Constrain panning to always remain inside pgA canvas.

* refactor(unsynced-panning): Move panning and zooming logic to a class

* feat(unsynced-panning): Add ObservableViewport instance to sketch.js

* feat(unsynced-panning): Add zoom to center (fix #22, #21)